refactor: rename Tool to DarkModeToggle
